requirements.txt
flask
flask-login
flask-mysqldb
bcrypt
python-dotenv
google-api-python-client
authlib
requests
flask-mysql
flask-bcrypt
google-auth
google-auth-oauthlib
streamlit
pandas
numpy
sqlalchemy
psycopg2-binary
plotly

